feat: move dapp-provider to own package
