+++ core/provider-dapp/jest.config.ts
@@ -1,0 +1,13 @@
+// Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import type { Config } from 'jest'
+
+export default {
+    rootDir: 'src',
+    extensionsToTreatAsEsm: ['.ts'],
+    resolver: 'ts-jest-resolver',
+    transform: {
+        '^.+\\.(t|j)sx?$': '@swc/jest',
+    },
+} satisfies Config
