+++ core/provider-dapp/package.json
@@ -1,10 +1,10 @@
 {
-    "name": "@canton-network/core-splice-provider",
-    "version": "0.24.0",
+    "name": "@canton-network/core-provider-dapp",
+    "version": "0.1.0",
     "type": "module",
-    "description": "A JavaScript Splice Provider API (EIP-1193 compliant).",
+    "description": "A Splice Provider implementation for web dApps.",
     "license": "Apache-2.0",
-    "author": "Marc Juchli <marc.juchli@digitalasset.com>",
+    "author": "Alex Matson <alex.matson@digitalasset.com>",
     "packageManager": "yarn@4.9.4",
     "main": "./dist/index.cjs",
     "module": "./dist/index.js",
@@ -20,16 +20,27 @@
     "scripts": {
         "build": "tsup && tsc -p tsconfig.types.json",
         "dev": "tsup --watch --onSuccess \"tsc -p tsconfig.types.json\"",
+        "test": "jest",
         "flatpack": "yarn pack --out \"$FLATPACK_OUTDIR\"",
         "clean": "tsc -b --clean; rm -rf dist"
     },
     "devDependencies": {
+        "@jest/globals": "^30.2.0",
+        "@swc/core": "^1.15.11",
+        "@swc/jest": "^0.2.39",
+        "@types/jest": "^30.0.0",
+        "jest": "^30.2.0",
+        "ts-jest": "^29.4.6",
+        "ts-jest-resolver": "^2.0.1",
         "tsup": "^8.5.1",
         "typescript": "^5.9.3"
     },
     "dependencies": {
         "@canton-network/core-rpc-transport": "workspace:^",
-        "@canton-network/core-types": "workspace:^"
+        "@canton-network/core-splice-provider": "workspace:^",
+        "@canton-network/core-types": "workspace:^",
+        "@canton-network/core-wallet-dapp-remote-rpc-client": "workspace:^",
+        "@canton-network/core-wallet-dapp-rpc-client": "workspace:^"
     },
     "files": [
         "dist/**"
@@ -40,6 +51,6 @@
     "repository": {
         "type": "git",
         "url": "git+https://github.com/hyperledger-labs/splice-wallet-kernel.git",
-        "directory": "core/splice-provider"
+        "directory": "core/provider-dapp"
     }
 }
