+++ core/provider-dapp/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../tsconfig.web.json",
+    "compilerOptions": {
+        "rootDir": "./src",
+        "outDir": "./dist"
+    },
+    "include": ["src"]
+}
