+++ core/provider-dapp/tsconfig.types.json
@@ -1,0 +1,15 @@
+{
+    "extends": "../../tsconfig.web.json",
+    "compilerOptions": {
+        "emitDeclarationOnly": true,
+        "declaration": true,
+        "declarationMap": true,
+        "rootDir": "./src",
+        "outDir": "./dist",
+        "moduleResolution": "bundler",
+        "composite": false,
+        "noEmit": false
+    },
+    "include": ["src/**/*.ts", "src/**/*.d.ts"],
+    "exclude": ["**/*.test.*", "**/*.spec.*", "**/__tests__/**"]
+}
