+++ sdk/dapp-sdk/src/error.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { DappProvider } from '@canton-network/core-splice-provider'
+import { DappProvider } from '@canton-network/core-provider-dapp'
 
 export enum ErrorCode {
     ProviderNotFound,
